requirements.txt
fastapi
uvicorn
scikit-learn
numpy
pandas
python-dotenv
jinja2
jupyter
ipykernel
jupyterlab
google-genai
seaborn
sentence-transformers
matplotlib
joblib
pydantic
pydantic-settings
langchain
langchain-community
chromadb
streamlit
langchain-chroma
